mach-imx: imx8m: soc: Add an empty entry for OPTEE at MMU table
When OPTEE is enabled, it splits the physical memory, needing an extra
entry at the MMU table for that. Variscite's boards gets its DRAM size
dinamically from the EEPROM, not defining the PHYS_SDRAM_2_SIZE macro.
But the definition of this macro is needed to include the DRAM2 entry
to the MMU table, and if it is not defined and the board has more than
3 GB of DRAM, the entry that was supposed to be used for the OPTEE is
used for the DRAM2, resulting in the board hanging when booting.

The DRAM1 and DRAM2 entries's starting values in the MMU table are
overwritten according to Variscite EEPROM by the enable_caches()
function:

https://github.com/varigit/uboot-imx/blob/lf_v2021.04_var03/arch/arm/mach-imx/imx8m/soc.c#L196-#L207

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>
(cherry picked from commit f5ddf1f684a908b3b696590dfc32c2f34d957833)
(cherry picked from commit 8cb07fc210b19976aee80c15a589c3a632263c73)
Signed-off-by: Stefano Radaelli <stefano.r@variscite.com>