imx8mp_var_dart: set default nfsroot
Derived from: 5c2d782

Setting a default NFS root directory allows to TFTP / NFS boot without
any need to modify the environment.

This allows to boot SOMs that come directly from factory. This could be
used for all:
- development
- automated testing with CI
- flashing at customer site

For this to work, the following IT setup must work:
- DHCP server reports a bootstrap server. This will automatically be
stored in the serverip field
- The reported bootstrap server provides a TFTP server from which the
default DTB files and the kernel image can be downloaded
- The reported bootstrap server provides an NFS server and is exporting
/srv/nfs/<machine-name>/rootfs

Signed-off-by: Bruno Bonaldi Teixeira <bruno.t@variscite.com>