configs: variscite: Center U-Boot splash logo
Center U-Boot splash logo for Variscite imx6ul, imx8qxp, imx95, and imx7d
based boards. Set the splashpos environment variable to m,m to center the
logo on the screen during U-Boot splash display. This ensures the logo is
displayed in the middle of the screen regardless of the panel resolution.

Signed-off-by: Marius Bagassien <marius.b@variscite.com>