imx8mp_var_dart: Remove GPIO expander 3 node for DT8MCustomBoard
The GPIO expander 3 node was added to the imx8mp-var-dart-dt8mcustomboard
device tree to support the new Sonata carrier board. However, this causes
the following error in U-Boot, when using the DT8MCustomBoard:

  pca953x gpio@22: Error reading output register

To fix this, remove the node from the device tree at runtime using
board_fix_fdt() when the DT8MCustomBoard carrier is detected.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>