imx95_var_dart: power on m7 for dart-mx95-m7/m7deb cfgs to kick m7 in A55.
OEI power off M7 if no m7 image in flash.bin. So need power on M7 in uboot
stage when need AP to manage M7.

To make power_on_m7 successfully, the configuration name passed as an argument
must match the configuration name used to build SM.
If the match fails, the M7 will not be powered on and running M7 firmwares
from U-Boot fail.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>