imx95_var_dart: add m7 suffix to load kernel dtb
With the i.MX95 SOC, the resources must be assigned to the various cores through the
SM configuration file.
When the SM is built using Cortex M7 configurations such as dart-mx95-m7 or dart-mx95-m7deb,
if the kernel attempts to use a resource assigned to the Cortex M7, a crash occurs.
Therefore, the kernel must load the device tree, which disables the resources used by the
Cortex M7 on the kernel (Cortex A55) side.
To avoid this problem, findfdt has been updated accordingly.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>