Revert "imx8mp-var-dart: Enable pull-up on ENET1_RGMII_RD3 for MXL MDIO strap"
This reverts commit 9bab4e2f446e.

Commit 9bab4e2f446e enabled the internal pull-up resistor on
ENET1_RGMII_RD3 to strap the MXL86110 MDIO address.

For the Sonata, equipped with a MXL86110, this pulls PHYAD0 high,
setting the PHY address to 1. However, on the DT8MCB (with AR8033
or ADIN1300), RD3 maps to PHYAD3. Since PHYAD0 already has an
external pull-up, this results in PHYAD3 being pulled high as
well, setting the address to 9 instead of 1. The PHY is then
not detected/configured properly, leading to:
  Could not get PHY for ethernet@30bf0000: addr 1

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>