configs: mx7dvar_som: Fix Cortex-M4 image loading
Due to the LMB (Logical Memory Block) check, loading the Cortex-M4 image
directly into TCM or OCRAM is prohibited.

This results in the following error:
```
Normal Boot
Hit any key to stop autoboot:  0
switch to partitions #0, OK
mmc0 is current device
** Reading file would overwrite reserved memory **
Failed to load '/hello_world.bin'
```

To fix this, load the image into the scratch area and copy it from there
to the final destination.

This approach is also transparent for Cortex-M4 images located in DDR.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>