imx95_var_dart.c: Reset the ethernet phy only if exists
Read the data from the EEPROM before calling netc_init() so the SoM's
features, in particular the presentense of the second ethernet, will
be available for netc_init().

This will set the voltage of the external phy to 1.8V (if the second
ethernet exists), otherwise leave it on 3.3V (default).

Signed-off-by: Yair Ben Avraham <yair.b@variscite.com>